Use caml_executable_name in header.c
